backend/requirements.txt
accelerate==0.30.1
adapters==0.2.0
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1716290248287/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1708355285029/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1695989838345/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
deprecation @ file:///home/conda/feedstock_root/build_artifacts/deprecation_1589881437857/work
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
fsspec==2024.3.1
gotrue @ file:///home/conda/feedstock_root/build_artifacts/gotrue_1711499799661/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
httpcore @ file:///home/conda/feedstock_root/build_artifacts/httpcore_1711596990900/work
httptools==0.6.1
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1701072546356/work
huggingface-hub==0.23.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.2.1
langchain-openai==0.1.7
langsmith==0.1.60
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pinecone-client==4.1.0
postgrest @ file:///home/conda/feedstock_root/build_artifacts/postgrest_1714493537416/work
psutil==5.9.8
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1713905669586/work
pydantic_core @ file:///Users/runner/miniforge3/conda-bld/pydantic-core_1713861857318/work
Pygments==2.18.0
PyJWT==2.8.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
realtime @ file:///home/conda/feedstock_root/build_artifacts/realtime_1713017269667/work
regex==2024.5.10
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1716354486713/work
rich==13.7.1
safetensors==0.4.3
setuptools==69.5.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1708952932303/work
starlette==0.37.2
storage3 @ file:///home/conda/feedstock_root/build_artifacts/storage3_1711251992635/work
StrEnum @ file:///home/conda/feedstock_root/build_artifacts/strenum_1688131856999/work
supabase @ file:///home/conda/feedstock_root/build_artifacts/supabase_1714608728706/work
supafunc @ file:///home/conda/feedstock_root/build_artifacts/supafunc_1711243718142/work
sympy==1.12
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.4
transformers==4.39.3
typer==0.12.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
ujson==5.10.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1708239446578/work
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets @ file:///Users/runner/miniforge3/conda-bld/websockets_1697914710084/work
wheel==0.43.0
